+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025/03/17
+
+### `Added`
+
+- `earliest` transformation: finds earliest date among metadata [PR #9](https://github.com/phac-nml/metadatatransformation/pull/9)
+- `populate` transformation: populates an output column with a specific value [PR #10](https://github.com/phac-nml/metadatatransformation/pull/10)
+
+### `Fixed`
+
+- A crash when calculating age if an entire date column contained only floats.
+
 ## [1.0.0] - 2025/03/07
 
 Initial release of phac-nml/metadatatransformation.
@@ -19,3 +30,4 @@ Initial release of phac-nml/metadatatransformation.
 ### `Deprecated`
 
 [1.0.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.0.0
+[1.1.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.1.0